index/returnPayload: Fix undefined reference t
Following a prior refactor, the `returnPayload` function included
references to the variable `t` to access its contents, but that variable
is not defined in this scope.

This will result in the following error if used in TODO/DOING mode and a
LOGBOOK is available.

```
Uncaught (in promise) ReferenceError: t is not defined
    at o (index.js:86)
    at index.js:98
```

Fix the error by accessing the provided content variable directly.